refactor: linter
