+++ src/safeds/data/image/containers/_image_list.py
@@ -18,7 +18,8 @@ from safeds.data.image.containers._image import Image
 if TYPE_CHECKING:
     from collections.abc import Sequence
 
-    from safeds.data.image.containers import _MultiSizeImageList, _SingleSizeImageList
+    from safeds.data.image.containers._multi_size_image_list import _MultiSizeImageList
+    from safeds.data.image.containers._single_size_image_list import _SingleSizeImageList
 
 
 class ImageList(metaclass=ABCMeta):
